Add MIT License to the repository to allow open-source usage, modification, and distribution
This commit includes the addition of the MIT License, granting permission to anyone who obtains a copy of the code to use, copy, modify, merge, publish, distribute, sublicense, and/or sell the software. The license also disclaims any warranties and limits liability. This will make the repository freely available for anyone to contribute to and use while ensuring that proper credit is given to the original author.